+++ recpack/splitters/__init__.py
@@ -37,7 +37,7 @@ Scenarios
     scenarios.Timed
     scenarios.StrongGeneralizationTimed
     scenarios.StrongGeneralizationTimedMostRecent
-    scenarios.NextItemPrediction
+    scenarios.LastItemPrediction
 
 Splitters
 ------------