+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * Added CosmeticsShop for https://www.kaggle.com/mkechinov/ecommerce-events-history-in-cosmetics-shop
     * Added RetailRocket for https://www.kaggle.com/retailrocket/ecommerce-dataset
     * Added parameters to dummy dataset to define the output expectations.
+* __algorithms.baseline__
+    * Added boolean parameter `use_only_interacted_items` to Random baseline, selects if all items should be used, or only those interacted with in the training dataset.
+* __splitters.scenarios__
+    * Added parameter `n_most_recent` to `NextItemPrediction` class to limit test_in data to only the N most recent interactions of each user.
 
 * __preprocessing.filters__
     * Added MaxItemsPerUser filter to remove users with extreme amounts of interactions from a dataframe.
@@ -22,7 +26,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * Refactored WeightedMatrixFactorization
         * It now uses PyTorch operations instead of NumPy (for GPU speedups).
         * It now processes batches of users and items instead of individual users and items.
+* __splitters.scenarios__
+    * Renamed `NextItemPrediction` to `LastItemPrediction` class, kept `NextItemPrediction` as a copy with deprecation warning. 
 
+### Bugfixes
+* __splitters.scenarios__
+    * Fixed bug in `NextItemPrediction` scenario:
+        * if validation was specified, test_in data contained 1 too few interactions
 ## [0.2.1] - ![](https://img.shields.io/date/1634291547.svg?label=2021-10-15)
 ### Dependency Update
 * Removed dependency on numba